Input: added check on UTM requirement
To guarantee correct application of `d` in the hierarchical clustering, location coords must be in UTM so that underlying distance matrix is in meters. Sure, this likely already done by us on the standardization app, but it'd make this app strongly dependent on an optional setting of an upstream app, which is a bit clunky. This fixes it, while still acknowledging if the option was used in the standardization app